+++ src/k8s_sandbox/_pod/pod.py
@@ -30,6 +30,7 @@ class Pod:
         stdin: str | bytes | None,
         cwd: str | None,
         env: dict[str, str],
+        user: str | None,
         timeout: int | None,
     ) -> ExecResult[str]:
         """
@@ -68,6 +69,9 @@ class Pod:
             directory does not exist, an unsuccessful ExecResult will be returned and
             cmd will not be run.
           env (dict[str, str]): The environment variables to set before running cmd.
+          user (str | None): The user to run the command as. If None, the default user
+            for the pod will be used. The container must be running as root to run as a
+            different user and the runuser command must be available in the container.
           timeout (int | None): The optional timeout for cmd to complete in. Defaults to
             no timeout. If provided, SIGTERM will be sent to cmd once the timeout has
             elapsed. This is enforced by the `timeout` command on the pod. This will not
@@ -75,7 +79,7 @@ class Pod:
         """
         executor = ExecuteOperation(self.info)
         result = await self._run_async(
-            lambda: executor.exec(cmd, stdin, cwd, env, timeout)
+            lambda: executor.exec(cmd, stdin, cwd, env, user, timeout)
         )
         return result